// NhapKhoApp/src/generated/models/Htn1423_onmuachitietsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Htn1423_onmuachitietsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Htn1423_onmuachitietsstatecode = keyof typeof Htn1423_onmuachitietsstatecode;
export const Htn1423_onmuachitietsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Htn1423_onmuachitietsstatuscode = keyof typeof Htn1423_onmuachitietsstatuscode;

export interface Htn1423_onmuachitietsBase {
  htn1423_name: string;
  htn1423_ongia: string;
  "htn1423_onhang@odata.bind": string;
  htn1423_onmuachitietid: string;
  htn1423_soluong: string;
  "htn1423_Tensanpham@odata.bind": string;
  importsequencenumber?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Htn1423_onmuachitietsstatecode;
  statuscode?: Htn1423_onmuachitietsstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Htn1423_onmuachitiets extends Htn1423_onmuachitietsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  htn1423_onhangname?: string;
  htn1423_tensanphamname?: string;
  htn1423_thanhtien?: string;
  htn1423_vat?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  htn1423_onhang?: object;
  _htn1423_onhang_value?: string;
  htn1423_tensanpham?: object;
  _htn1423_tensanpham_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}